Revert "sys/src/9/vt4: fix prototype of ainc and adec"
This reverts commit 8a9afd83b47cfe7cb04877a5e5fe69ada05590e9.

+++ sys/src/9/vt4/fns.h
@@ -8,8 +8,8 @@ typedef struct Ether Ether;
 #pragma incomplete Dma
 
 void	addconf(char*, char*);
-long	adec(long*);
-long	ainc(long*);
+int	adec(long*);
+int	ainc(long*);
 void	archetherspeed(int, int);
 void	archinit(void);
 uint	archmiiport(int);
